Add player jump system with parabolic arc and catch height gating
- Add JUMPING status (bit 5), jumpFrame/jumpHeight fields, and startJump() to Player; JUMP_FRAMES=30
- runAI() processes jump arc each frame (parabolic peak at stats.jump px), locks direction until landing
- Renderer subtracts jumpHeight from draw Y for player and carried football
- Catching::catchBall gates on live jumpHeight instead of static jump stat
- Change PlayerStats.jump from float to int; update all roster values

+++ arm9/source/players/player_stats.h
@@ -12,6 +12,6 @@ struct PlayerStats {
     int   block         = 0;
     int   blockShed     = 0;
     float throwStrength = 0.0f;
-    float jump         = 0.0f;
+    int   jump         = 0;
     int   catching     = 0;
 };
